Adds basic HomeActivity and WelcomeFragment
Implements a basic HomeActivity to serve as the main screen after login.
Adds a WelcomeFragment to display a welcome message and a "Get Started" button.
The "Get Started" button is currently non-functional and will be implemented later.
The HomeActivity uses a FrameLayout to host the WelcomeFragment.
Includes the required layout files and drawables.